Struct rusoto_config::ConfigServiceClient

pub struct ConfigServiceClient { /* fields omitted */ }

A client for the Config Service API.

Methods

impl ConfigServiceClient

pub fn new(region: Region) -> ConfigServiceClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> ConfigServiceClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,