Struct rusoto_config::ConfigurationItem

pub struct ConfigurationItem {
    pub account_id: Option<String>,
    pub arn: Option<String>,
    pub availability_zone: Option<String>,
    pub aws_region: Option<String>,
    pub configuration: Option<String>,
    pub configuration_item_capture_time: Option<f64>,
    pub configuration_item_md5_hash: Option<String>,
    pub configuration_item_status: Option<String>,
    pub configuration_state_id: Option<String>,
    pub related_events: Option<Vec<String>>,
    pub relationships: Option<Vec<Relationship>>,
    pub resource_creation_time: Option<f64>,
    pub resource_id: Option<String>,
    pub resource_name: Option<String>,
    pub resource_type: Option<String>,
    pub supplementary_configuration: Option<HashMap<String, String>>,
    pub tags: Option<HashMap<String, String>>,
    pub version: Option<String>,
}

A list that contains detailed configurations of a specified resource.

Fields
account_id: Option<String>

The 12-digit AWS account ID associated with the resource.

arn: Option<String>

The Amazon Resource Name (ARN) of the resource.

availability_zone: Option<String>

The Availability Zone associated with the resource.

aws_region: Option<String>

The region where the resource resides.

configuration: Option<String>

The description of the resource configuration.

configuration_item_capture_time: Option<f64>

The time when the configuration recording was initiated.

configuration_item_md5_hash: Option<String>

Unique MD5 hash that represents the configuration item's state.

You can use MD5 hash to compare the states of two or more configuration items that are associated with the same resource.

configuration_item_status: Option<String>

The configuration item status.

configuration_state_id: Option<String>

An identifier that indicates the ordering of the configuration items of a resource.

related_events: Option<Vec<String>>

A list of CloudTrail event IDs.

A populated field indicates that the current configuration was initiated by the events recorded in the CloudTrail log. For more information about CloudTrail, see What Is AWS CloudTrail.

An empty field indicates that the current configuration was not initiated by any event.

relationships: Option<Vec<Relationship>>

A list of related AWS resources.

resource_creation_time: Option<f64>

The time stamp when the resource was created.

resource_id: Option<String>

The ID of the resource (for example, sg-xxxxxx).

resource_name: Option<String>

The custom name of the resource, if available.

resource_type: Option<String>

The type of AWS resource.

supplementary_configuration: Option<HashMap<String, String>>

Configuration attributes that AWS Config returns for certain resource types to supplement the information returned for the configuration parameter.

tags: Option<HashMap<String, String>>

A mapping of key value tags associated with the resource.

version: Option<String>

The version number of the resource configuration.