Trait rusoto_config::ConfigService

pub trait ConfigService {
    fn batch_get_aggregate_resource_config(
        &self, 
        input: BatchGetAggregateResourceConfigRequest
    ) -> RusotoFuture<BatchGetAggregateResourceConfigResponse, BatchGetAggregateResourceConfigError>;

    fn batch_get_resource_config(
        &self, 
        input: BatchGetResourceConfigRequest
    ) -> RusotoFuture<BatchGetResourceConfigResponse, BatchGetResourceConfigError>;

    fn delete_aggregation_authorization(
        &self, 
        input: DeleteAggregationAuthorizationRequest
    ) -> RusotoFuture<(), DeleteAggregationAuthorizationError>;

    fn delete_config_rule(
        &self, 
        input: DeleteConfigRuleRequest
    ) -> RusotoFuture<(), DeleteConfigRuleError>;

    fn delete_configuration_aggregator(
        &self, 
        input: DeleteConfigurationAggregatorRequest
    ) -> RusotoFuture<(), DeleteConfigurationAggregatorError>;

    fn delete_configuration_recorder(
        &self, 
        input: DeleteConfigurationRecorderRequest
    ) -> RusotoFuture<(), DeleteConfigurationRecorderError>;

    fn delete_delivery_channel(
        &self, 
        input: DeleteDeliveryChannelRequest
    ) -> RusotoFuture<(), DeleteDeliveryChannelError>;

    fn delete_evaluation_results(
        &self, 
        input: DeleteEvaluationResultsRequest
    ) -> RusotoFuture<DeleteEvaluationResultsResponse, DeleteEvaluationResultsError>;

    fn delete_pending_aggregation_request(
        &self, 
        input: DeletePendingAggregationRequestRequest
    ) -> RusotoFuture<(), DeletePendingAggregationRequestError>;

    fn delete_remediation_configuration(
        &self, 
        input: DeleteRemediationConfigurationRequest
    ) -> RusotoFuture<DeleteRemediationConfigurationResponse, DeleteRemediationConfigurationError>;

    fn delete_retention_configuration(
        &self, 
        input: DeleteRetentionConfigurationRequest
    ) -> RusotoFuture<(), DeleteRetentionConfigurationError>;

    fn deliver_config_snapshot(
        &self, 
        input: DeliverConfigSnapshotRequest
    ) -> RusotoFuture<DeliverConfigSnapshotResponse, DeliverConfigSnapshotError>;

    fn describe_aggregate_compliance_by_config_rules(
        &self, 
        input: DescribeAggregateComplianceByConfigRulesRequest
    ) -> RusotoFuture<DescribeAggregateComplianceByConfigRulesResponse, DescribeAggregateComplianceByConfigRulesError>;

    fn describe_aggregation_authorizations(
        &self, 
        input: DescribeAggregationAuthorizationsRequest
    ) -> RusotoFuture<DescribeAggregationAuthorizationsResponse, DescribeAggregationAuthorizationsError>;

    fn describe_compliance_by_config_rule(
        &self, 
        input: DescribeComplianceByConfigRuleRequest
    ) -> RusotoFuture<DescribeComplianceByConfigRuleResponse, DescribeComplianceByConfigRuleError>;

    fn describe_compliance_by_resource(
        &self, 
        input: DescribeComplianceByResourceRequest
    ) -> RusotoFuture<DescribeComplianceByResourceResponse, DescribeComplianceByResourceError>;

    fn describe_config_rule_evaluation_status(
        &self, 
        input: DescribeConfigRuleEvaluationStatusRequest
    ) -> RusotoFuture<DescribeConfigRuleEvaluationStatusResponse, DescribeConfigRuleEvaluationStatusError>;

    fn describe_config_rules(
        &self, 
        input: DescribeConfigRulesRequest
    ) -> RusotoFuture<DescribeConfigRulesResponse, DescribeConfigRulesError>;

    fn describe_configuration_aggregator_sources_status(
        &self, 
        input: DescribeConfigurationAggregatorSourcesStatusRequest
    ) -> RusotoFuture<DescribeConfigurationAggregatorSourcesStatusResponse, DescribeConfigurationAggregatorSourcesStatusError>;

    fn describe_configuration_aggregators(
        &self, 
        input: DescribeConfigurationAggregatorsRequest
    ) -> RusotoFuture<DescribeConfigurationAggregatorsResponse, DescribeConfigurationAggregatorsError>;

    fn describe_configuration_recorder_status(
        &self, 
        input: DescribeConfigurationRecorderStatusRequest
    ) -> RusotoFuture<DescribeConfigurationRecorderStatusResponse, DescribeConfigurationRecorderStatusError>;

    fn describe_configuration_recorders(
        &self, 
        input: DescribeConfigurationRecordersRequest
    ) -> RusotoFuture<DescribeConfigurationRecordersResponse, DescribeConfigurationRecordersError>;

    fn describe_delivery_channel_status(
        &self, 
        input: DescribeDeliveryChannelStatusRequest
    ) -> RusotoFuture<DescribeDeliveryChannelStatusResponse, DescribeDeliveryChannelStatusError>;

    fn describe_delivery_channels(
        &self, 
        input: DescribeDeliveryChannelsRequest
    ) -> RusotoFuture<DescribeDeliveryChannelsResponse, DescribeDeliveryChannelsError>;

    fn describe_pending_aggregation_requests(
        &self, 
        input: DescribePendingAggregationRequestsRequest
    ) -> RusotoFuture<DescribePendingAggregationRequestsResponse, DescribePendingAggregationRequestsError>;

    fn describe_remediation_configurations(
        &self, 
        input: DescribeRemediationConfigurationsRequest
    ) -> RusotoFuture<DescribeRemediationConfigurationsResponse, DescribeRemediationConfigurationsError>;

    fn describe_remediation_execution_status(
        &self, 
        input: DescribeRemediationExecutionStatusRequest
    ) -> RusotoFuture<DescribeRemediationExecutionStatusResponse, DescribeRemediationExecutionStatusError>;

    fn describe_retention_configurations(
        &self, 
        input: DescribeRetentionConfigurationsRequest
    ) -> RusotoFuture<DescribeRetentionConfigurationsResponse, DescribeRetentionConfigurationsError>;

    fn get_aggregate_compliance_details_by_config_rule(
        &self, 
        input: GetAggregateComplianceDetailsByConfigRuleRequest
    ) -> RusotoFuture<GetAggregateComplianceDetailsByConfigRuleResponse, GetAggregateComplianceDetailsByConfigRuleError>;

    fn get_aggregate_config_rule_compliance_summary(
        &self, 
        input: GetAggregateConfigRuleComplianceSummaryRequest
    ) -> RusotoFuture<GetAggregateConfigRuleComplianceSummaryResponse, GetAggregateConfigRuleComplianceSummaryError>;

    fn get_aggregate_discovered_resource_counts(
        &self, 
        input: GetAggregateDiscoveredResourceCountsRequest
    ) -> RusotoFuture<GetAggregateDiscoveredResourceCountsResponse, GetAggregateDiscoveredResourceCountsError>;

    fn get_aggregate_resource_config(
        &self, 
        input: GetAggregateResourceConfigRequest
    ) -> RusotoFuture<GetAggregateResourceConfigResponse, GetAggregateResourceConfigError>;

    fn get_compliance_details_by_config_rule(
        &self, 
        input: GetComplianceDetailsByConfigRuleRequest
    ) -> RusotoFuture<GetComplianceDetailsByConfigRuleResponse, GetComplianceDetailsByConfigRuleError>;

    fn get_compliance_details_by_resource(
        &self, 
        input: GetComplianceDetailsByResourceRequest
    ) -> RusotoFuture<GetComplianceDetailsByResourceResponse, GetComplianceDetailsByResourceError>;

    fn get_compliance_summary_by_config_rule(
        &self
    ) -> RusotoFuture<GetComplianceSummaryByConfigRuleResponse, GetComplianceSummaryByConfigRuleError>;

    fn get_compliance_summary_by_resource_type(
        &self, 
        input: GetComplianceSummaryByResourceTypeRequest
    ) -> RusotoFuture<GetComplianceSummaryByResourceTypeResponse, GetComplianceSummaryByResourceTypeError>;

    fn get_discovered_resource_counts(
        &self, 
        input: GetDiscoveredResourceCountsRequest
    ) -> RusotoFuture<GetDiscoveredResourceCountsResponse, GetDiscoveredResourceCountsError>;

    fn get_resource_config_history(
        &self, 
        input: GetResourceConfigHistoryRequest
    ) -> RusotoFuture<GetResourceConfigHistoryResponse, GetResourceConfigHistoryError>;

    fn list_aggregate_discovered_resources(
        &self, 
        input: ListAggregateDiscoveredResourcesRequest
    ) -> RusotoFuture<ListAggregateDiscoveredResourcesResponse, ListAggregateDiscoveredResourcesError>;

    fn list_discovered_resources(
        &self, 
        input: ListDiscoveredResourcesRequest
    ) -> RusotoFuture<ListDiscoveredResourcesResponse, ListDiscoveredResourcesError>;

    fn list_tags_for_resource(
        &self, 
        input: ListTagsForResourceRequest
    ) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>;

    fn put_aggregation_authorization(
        &self, 
        input: PutAggregationAuthorizationRequest
    ) -> RusotoFuture<PutAggregationAuthorizationResponse, PutAggregationAuthorizationError>;

    fn put_config_rule(
        &self, 
        input: PutConfigRuleRequest
    ) -> RusotoFuture<(), PutConfigRuleError>;

    fn put_configuration_aggregator(
        &self, 
        input: PutConfigurationAggregatorRequest
    ) -> RusotoFuture<PutConfigurationAggregatorResponse, PutConfigurationAggregatorError>;

    fn put_configuration_recorder(
        &self, 
        input: PutConfigurationRecorderRequest
    ) -> RusotoFuture<(), PutConfigurationRecorderError>;

    fn put_delivery_channel(
        &self, 
        input: PutDeliveryChannelRequest
    ) -> RusotoFuture<(), PutDeliveryChannelError>;

    fn put_evaluations(
        &self, 
        input: PutEvaluationsRequest
    ) -> RusotoFuture<PutEvaluationsResponse, PutEvaluationsError>;

    fn put_remediation_configurations(
        &self, 
        input: PutRemediationConfigurationsRequest
    ) -> RusotoFuture<PutRemediationConfigurationsResponse, PutRemediationConfigurationsError>;

    fn put_retention_configuration(
        &self, 
        input: PutRetentionConfigurationRequest
    ) -> RusotoFuture<PutRetentionConfigurationResponse, PutRetentionConfigurationError>;

    fn select_resource_config(
        &self, 
        input: SelectResourceConfigRequest
    ) -> RusotoFuture<SelectResourceConfigResponse, SelectResourceConfigError>;

    fn start_config_rules_evaluation(
        &self, 
        input: StartConfigRulesEvaluationRequest
    ) -> RusotoFuture<StartConfigRulesEvaluationResponse, StartConfigRulesEvaluationError>;

    fn start_configuration_recorder(
        &self, 
        input: StartConfigurationRecorderRequest
    ) -> RusotoFuture<(), StartConfigurationRecorderError>;

    fn start_remediation_execution(
        &self, 
        input: StartRemediationExecutionRequest
    ) -> RusotoFuture<StartRemediationExecutionResponse, StartRemediationExecutionError>;

    fn stop_configuration_recorder(
        &self, 
        input: StopConfigurationRecorderRequest
    ) -> RusotoFuture<(), StopConfigurationRecorderError>;

    fn tag_resource(
        &self, 
        input: TagResourceRequest
    ) -> RusotoFuture<(), TagResourceError>;

    fn untag_resource(
        &self, 
        input: UntagResourceRequest
    ) -> RusotoFuture<(), UntagResourceError>;
}

Trait representing the capabilities of the Config Service API. Config Service clients implement this trait.

Required methods

fn batch_get_aggregate_resource_config(
&self,
input: BatchGetAggregateResourceConfigRequest
) -> RusotoFuture<BatchGetAggregateResourceConfigResponse, BatchGetAggregateResourceConfigError>

Returns the current configuration items for resources that are present in your AWS Config aggregator. The operation also returns a list of resources that are not processed in the current request. If there are no unprocessed resources, the operation returns an empty unprocessedResourceIdentifiers list.

The API does not return results for deleted resources.

The API does not return tags and relationships.

fn batch_get_resource_config(
&self,
input: BatchGetResourceConfigRequest
) -> RusotoFuture<BatchGetResourceConfigResponse, BatchGetResourceConfigError>

Returns the current configuration for one or more requested resources. The operation also returns a list of resources that are not processed in the current request. If there are no unprocessed resources, the operation returns an empty unprocessedResourceKeys list.

The API does not return results for deleted resources.

The API does not return any tags for the requested resources. This information is filtered out of the supplementaryConfiguration section of the API response.

fn delete_aggregation_authorization(
&self,
input: DeleteAggregationAuthorizationRequest
) -> RusotoFuture<(), DeleteAggregationAuthorizationError>

Deletes the authorization granted to the specified configuration aggregator account in a specified region.

fn delete_config_rule(
&self,
input: DeleteConfigRuleRequest
) -> RusotoFuture<(), DeleteConfigRuleError>

Deletes the specified AWS Config rule and all of its evaluation results.

AWS Config sets the state of a rule to DELETING until the deletion is complete. You cannot update a rule while it is in this state. If you make a PutConfigRule or DeleteConfigRule request for the rule, you will receive a ResourceInUseException.

You can check the state of a rule by using the DescribeConfigRules request.

fn delete_configuration_aggregator(
&self,
input: DeleteConfigurationAggregatorRequest
) -> RusotoFuture<(), DeleteConfigurationAggregatorError>

Deletes the specified configuration aggregator and the aggregated data associated with the aggregator.

fn delete_configuration_recorder(
&self,
input: DeleteConfigurationRecorderRequest
) -> RusotoFuture<(), DeleteConfigurationRecorderError>

Deletes the configuration recorder.

After the configuration recorder is deleted, AWS Config will not record resource configuration changes until you create a new configuration recorder.

This action does not delete the configuration information that was previously recorded. You will be able to access the previously recorded information by using the GetResourceConfigHistory action, but you will not be able to access this information in the AWS Config console until you create a new configuration recorder.

fn delete_delivery_channel(
&self,
input: DeleteDeliveryChannelRequest
) -> RusotoFuture<(), DeleteDeliveryChannelError>

Deletes the delivery channel.

Before you can delete the delivery channel, you must stop the configuration recorder by using the StopConfigurationRecorder action.

fn delete_evaluation_results(
&self,
input: DeleteEvaluationResultsRequest
) -> RusotoFuture<DeleteEvaluationResultsResponse, DeleteEvaluationResultsError>

Deletes the evaluation results for the specified AWS Config rule. You can specify one AWS Config rule per request. After you delete the evaluation results, you can call the StartConfigRulesEvaluation API to start evaluating your AWS resources against the rule.

fn delete_pending_aggregation_request(
&self,
input: DeletePendingAggregationRequestRequest
) -> RusotoFuture<(), DeletePendingAggregationRequestError>

Deletes pending authorization requests for a specified aggregator account in a specified region.

fn delete_remediation_configuration(
&self,
input: DeleteRemediationConfigurationRequest
) -> RusotoFuture<DeleteRemediationConfigurationResponse, DeleteRemediationConfigurationError>

Deletes the remediation configuration.

fn delete_retention_configuration(
&self,
input: DeleteRetentionConfigurationRequest
) -> RusotoFuture<(), DeleteRetentionConfigurationError>

Deletes the retention configuration.

fn deliver_config_snapshot(
&self,
input: DeliverConfigSnapshotRequest
) -> RusotoFuture<DeliverConfigSnapshotResponse, DeliverConfigSnapshotError>

Schedules delivery of a configuration snapshot to the Amazon S3 bucket in the specified delivery channel. After the delivery has started, AWS Config sends the following notifications using an Amazon SNS topic that you have specified.

Notification of the start of the delivery.

Notification of the completion of the delivery, if the delivery was successfully completed.

Notification of delivery failure, if the delivery failed.

fn describe_aggregate_compliance_by_config_rules(
&self,
input: DescribeAggregateComplianceByConfigRulesRequest
) -> RusotoFuture<DescribeAggregateComplianceByConfigRulesResponse, DescribeAggregateComplianceByConfigRulesError>

Returns a list of compliant and noncompliant rules with the number of resources for compliant and noncompliant rules.

The results can return an empty result page, but if you have a nextToken, the results are displayed on the next page.

fn describe_aggregation_authorizations(
&self,
input: DescribeAggregationAuthorizationsRequest
) -> RusotoFuture<DescribeAggregationAuthorizationsResponse, DescribeAggregationAuthorizationsError>

Returns a list of authorizations granted to various aggregator accounts and regions.

fn describe_compliance_by_config_rule(
&self,
input: DescribeComplianceByConfigRuleRequest
) -> RusotoFuture<DescribeComplianceByConfigRuleResponse, DescribeComplianceByConfigRuleError>

Indicates whether the specified AWS Config rules are compliant. If a rule is noncompliant, this action returns the number of AWS resources that do not comply with the rule.

A rule is compliant if all of the evaluated resources comply with it. It is noncompliant if any of these resources do not comply.

If AWS Config has no current evaluation results for the rule, it returns INSUFFICIENTDATA. This result might indicate one of the following conditions:

AWS Config has never invoked an evaluation for the rule. To check whether it has, use the DescribeConfigRuleEvaluationStatus action to get the LastSuccessfulInvocationTime and LastFailedInvocationTime.

The rule's AWS Lambda function is failing to send evaluation results to AWS Config. Verify that the role you assigned to your configuration recorder includes the config:PutEvaluations permission. If the rule is a custom rule, verify that the AWS Lambda execution role includes the config:PutEvaluations permission.

The rule's AWS Lambda function has returned NOTAPPLICABLE for all evaluation results. This can occur if the resources were deleted or removed from the rule's scope.

fn describe_compliance_by_resource(
&self,
input: DescribeComplianceByResourceRequest
) -> RusotoFuture<DescribeComplianceByResourceResponse, DescribeComplianceByResourceError>

Indicates whether the specified AWS resources are compliant. If a resource is noncompliant, this action returns the number of AWS Config rules that the resource does not comply with.

A resource is compliant if it complies with all the AWS Config rules that evaluate it. It is noncompliant if it does not comply with one or more of these rules.

If AWS Config has no current evaluation results for the resource, it returns INSUFFICIENTDATA. This result might indicate one of the following conditions about the rules that evaluate the resource:

AWS Config has never invoked an evaluation for the rule. To check whether it has, use the DescribeConfigRuleEvaluationStatus action to get the LastSuccessfulInvocationTime and LastFailedInvocationTime.

The rule's AWS Lambda function is failing to send evaluation results to AWS Config. Verify that the role that you assigned to your configuration recorder includes the config:PutEvaluations permission. If the rule is a custom rule, verify that the AWS Lambda execution role includes the config:PutEvaluations permission.

The rule's AWS Lambda function has returned NOTAPPLICABLE for all evaluation results. This can occur if the resources were deleted or removed from the rule's scope.

fn describe_config_rule_evaluation_status(
&self,
input: DescribeConfigRuleEvaluationStatusRequest
) -> RusotoFuture<DescribeConfigRuleEvaluationStatusResponse, DescribeConfigRuleEvaluationStatusError>

Returns status information for each of your AWS managed Config rules. The status includes information such as the last time AWS Config invoked the rule, the last time AWS Config failed to invoke the rule, and the related error for the last failure.

fn describe_config_rules(
&self,
input: DescribeConfigRulesRequest
) -> RusotoFuture<DescribeConfigRulesResponse, DescribeConfigRulesError>

Returns details about your AWS Config rules.

fn describe_configuration_aggregator_sources_status(
&self,
input: DescribeConfigurationAggregatorSourcesStatusRequest
) -> RusotoFuture<DescribeConfigurationAggregatorSourcesStatusResponse, DescribeConfigurationAggregatorSourcesStatusError>

Returns status information for sources within an aggregator. The status includes information about the last time AWS Config verified authorization between the source account and an aggregator account. In case of a failure, the status contains the related error code or message.

fn describe_configuration_aggregators(
&self,
input: DescribeConfigurationAggregatorsRequest
) -> RusotoFuture<DescribeConfigurationAggregatorsResponse, DescribeConfigurationAggregatorsError>

Returns the details of one or more configuration aggregators. If the configuration aggregator is not specified, this action returns the details for all the configuration aggregators associated with the account.

fn describe_configuration_recorder_status(
&self,
input: DescribeConfigurationRecorderStatusRequest
) -> RusotoFuture<DescribeConfigurationRecorderStatusResponse, DescribeConfigurationRecorderStatusError>

Returns the current status of the specified configuration recorder. If a configuration recorder is not specified, this action returns the status of all configuration recorders associated with the account.

Currently, you can specify only one configuration recorder per region in your account.

fn describe_configuration_recorders(
&self,
input: DescribeConfigurationRecordersRequest
) -> RusotoFuture<DescribeConfigurationRecordersResponse, DescribeConfigurationRecordersError>

Returns the details for the specified configuration recorders. If the configuration recorder is not specified, this action returns the details for all configuration recorders associated with the account.

Currently, you can specify only one configuration recorder per region in your account.

fn describe_delivery_channel_status(
&self,
input: DescribeDeliveryChannelStatusRequest
) -> RusotoFuture<DescribeDeliveryChannelStatusResponse, DescribeDeliveryChannelStatusError>

Returns the current status of the specified delivery channel. If a delivery channel is not specified, this action returns the current status of all delivery channels associated with the account.

Currently, you can specify only one delivery channel per region in your account.

fn describe_delivery_channels(
&self,
input: DescribeDeliveryChannelsRequest
) -> RusotoFuture<DescribeDeliveryChannelsResponse, DescribeDeliveryChannelsError>

Returns details about the specified delivery channel. If a delivery channel is not specified, this action returns the details of all delivery channels associated with the account.

Currently, you can specify only one delivery channel per region in your account.

fn describe_pending_aggregation_requests(
&self,
input: DescribePendingAggregationRequestsRequest
) -> RusotoFuture<DescribePendingAggregationRequestsResponse, DescribePendingAggregationRequestsError>

Returns a list of all pending aggregation requests.

fn describe_remediation_configurations(
&self,
input: DescribeRemediationConfigurationsRequest
) -> RusotoFuture<DescribeRemediationConfigurationsResponse, DescribeRemediationConfigurationsError>

Returns the details of one or more remediation configurations.

fn describe_remediation_execution_status(
&self,
input: DescribeRemediationExecutionStatusRequest
) -> RusotoFuture<DescribeRemediationExecutionStatusResponse, DescribeRemediationExecutionStatusError>

Provides a detailed view of a Remediation Execution for a set of resources including state, timestamps for when steps for the remediation execution occur, and any error messages for steps that have failed. When you specify the limit and the next token, you receive a paginated response.

fn describe_retention_configurations(
&self,
input: DescribeRetentionConfigurationsRequest
) -> RusotoFuture<DescribeRetentionConfigurationsResponse, DescribeRetentionConfigurationsError>

Returns the details of one or more retention configurations. If the retention configuration name is not specified, this action returns the details for all the retention configurations for that account.

Currently, AWS Config supports only one retention configuration per region in your account.

fn get_aggregate_compliance_details_by_config_rule(
&self,
input: GetAggregateComplianceDetailsByConfigRuleRequest
) -> RusotoFuture<GetAggregateComplianceDetailsByConfigRuleResponse, GetAggregateComplianceDetailsByConfigRuleError>

Returns the evaluation results for the specified AWS Config rule for a specific resource in a rule. The results indicate which AWS resources were evaluated by the rule, when each resource was last evaluated, and whether each resource complies with the rule.

The results can return an empty result page. But if you have a nextToken, the results are displayed on the next page.

fn get_aggregate_config_rule_compliance_summary(
&self,
input: GetAggregateConfigRuleComplianceSummaryRequest
) -> RusotoFuture<GetAggregateConfigRuleComplianceSummaryResponse, GetAggregateConfigRuleComplianceSummaryError>

Returns the number of compliant and noncompliant rules for one or more accounts and regions in an aggregator.

The results can return an empty result page, but if you have a nextToken, the results are displayed on the next page.

fn get_aggregate_discovered_resource_counts(
&self,
input: GetAggregateDiscoveredResourceCountsRequest
) -> RusotoFuture<GetAggregateDiscoveredResourceCountsResponse, GetAggregateDiscoveredResourceCountsError>

Returns the resource counts across accounts and regions that are present in your AWS Config aggregator. You can request the resource counts by providing filters and GroupByKey.

For example, if the input contains accountID 12345678910 and region us-east-1 in filters, the API returns the count of resources in account ID 12345678910 and region us-east-1. If the input contains ACCOUNT_ID as a GroupByKey, the API returns resource counts for all source accounts that are present in your aggregator.

fn get_aggregate_resource_config(
&self,
input: GetAggregateResourceConfigRequest
) -> RusotoFuture<GetAggregateResourceConfigResponse, GetAggregateResourceConfigError>

Returns configuration item that is aggregated for your specific resource in a specific source account and region.

fn get_compliance_details_by_config_rule(
&self,
input: GetComplianceDetailsByConfigRuleRequest
) -> RusotoFuture<GetComplianceDetailsByConfigRuleResponse, GetComplianceDetailsByConfigRuleError>

Returns the evaluation results for the specified AWS Config rule. The results indicate which AWS resources were evaluated by the rule, when each resource was last evaluated, and whether each resource complies with the rule.

fn get_compliance_details_by_resource(
&self,
input: GetComplianceDetailsByResourceRequest
) -> RusotoFuture<GetComplianceDetailsByResourceResponse, GetComplianceDetailsByResourceError>

Returns the evaluation results for the specified AWS resource. The results indicate which AWS Config rules were used to evaluate the resource, when each rule was last used, and whether the resource complies with each rule.

fn get_compliance_summary_by_config_rule(
&self
) -> RusotoFuture<GetComplianceSummaryByConfigRuleResponse, GetComplianceSummaryByConfigRuleError>

Returns the number of AWS Config rules that are compliant and noncompliant, up to a maximum of 25 for each.

fn get_compliance_summary_by_resource_type(
&self,
input: GetComplianceSummaryByResourceTypeRequest
) -> RusotoFuture<GetComplianceSummaryByResourceTypeResponse, GetComplianceSummaryByResourceTypeError>

Returns the number of resources that are compliant and the number that are noncompliant. You can specify one or more resource types to get these numbers for each resource type. The maximum number returned is 100.

fn get_discovered_resource_counts(
&self,
input: GetDiscoveredResourceCountsRequest
) -> RusotoFuture<GetDiscoveredResourceCountsResponse, GetDiscoveredResourceCountsError>

Returns the resource types, the number of each resource type, and the total number of resources that AWS Config is recording in this region for your AWS account.

Example

AWS Config is recording three resource types in the US East (Ohio) Region for your account: 25 EC2 instances, 20 IAM users, and 15 S3 buckets.

You make a call to the GetDiscoveredResourceCounts action and specify that you want all resource types.

AWS Config returns the following:

The resource types (EC2 instances, IAM users, and S3 buckets).

The number of each resource type (25, 20, and 15).

The total number of all resources (60).

The response is paginated. By default, AWS Config lists 100 ResourceCount objects on each page. You can customize this number with the limit parameter. The response includes a nextToken string. To get the next page of results, run the request again and specify the string for the nextToken parameter.

If you make a call to the GetDiscoveredResourceCounts action, you might not immediately receive resource counts in the following situations:

You are a new AWS Config customer.

You just enabled resource recording.

It might take a few minutes for AWS Config to record and count your resources. Wait a few minutes and then retry the GetDiscoveredResourceCounts action.

fn get_resource_config_history(
&self,
input: GetResourceConfigHistoryRequest
) -> RusotoFuture<GetResourceConfigHistoryResponse, GetResourceConfigHistoryError>

Returns a list of configuration items for the specified resource. The list contains details about each state of the resource during the specified time interval. If you specified a retention period to retain your ConfigurationItems between a minimum of 30 days and a maximum of 7 years (2557 days), AWS Config returns the ConfigurationItems for the specified retention period.

The response is paginated. By default, AWS Config returns a limit of 10 configuration items per page. You can customize this number with the limit parameter. The response includes a nextToken string. To get the next page of results, run the request again and specify the string for the nextToken parameter.

Each call to the API is limited to span a duration of seven days. It is likely that the number of records returned is smaller than the specified limit. In such cases, you can make another call, using the nextToken.

fn list_aggregate_discovered_resources(
&self,
input: ListAggregateDiscoveredResourcesRequest
) -> RusotoFuture<ListAggregateDiscoveredResourcesResponse, ListAggregateDiscoveredResourcesError>

Accepts a resource type and returns a list of resource identifiers that are aggregated for a specific resource type across accounts and regions. A resource identifier includes the resource type, ID, (if available) the custom resource name, source account, and source region. You can narrow the results to include only resources that have specific resource IDs, or a resource name, or source account ID, or source region.

For example, if the input consists of accountID 12345678910 and the region is us-east-1 for resource type AWS::EC2::Instance then the API returns all the EC2 instance identifiers of accountID 12345678910 and region us-east-1.

fn list_discovered_resources(
&self,
input: ListDiscoveredResourcesRequest
) -> RusotoFuture<ListDiscoveredResourcesResponse, ListDiscoveredResourcesError>

Accepts a resource type and returns a list of resource identifiers for the resources of that type. A resource identifier includes the resource type, ID, and (if available) the custom resource name. The results consist of resources that AWS Config has discovered, including those that AWS Config is not currently recording. You can narrow the results to include only resources that have specific resource IDs or a resource name.

You can specify either resource IDs or a resource name, but not both, in the same request.

The response is paginated. By default, AWS Config lists 100 resource identifiers on each page. You can customize this number with the limit parameter. The response includes a nextToken string. To get the next page of results, run the request again and specify the string for the nextToken parameter.

fn list_tags_for_resource(
&self,
input: ListTagsForResourceRequest
) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>

List the tags for AWS Config resource.

fn put_aggregation_authorization(
&self,
input: PutAggregationAuthorizationRequest
) -> RusotoFuture<PutAggregationAuthorizationResponse, PutAggregationAuthorizationError>

Authorizes the aggregator account and region to collect data from the source account and region.

fn put_config_rule(
&self,
input: PutConfigRuleRequest
) -> RusotoFuture<(), PutConfigRuleError>

Adds or updates an AWS Config rule for evaluating whether your AWS resources comply with your desired configurations.

You can use this action for custom AWS Config rules and AWS managed Config rules. A custom AWS Config rule is a rule that you develop and maintain. An AWS managed Config rule is a customizable, predefined rule that AWS Config provides.

If you are adding a new custom AWS Config rule, you must first create the AWS Lambda function that the rule invokes to evaluate your resources. When you use the PutConfigRule action to add the rule to AWS Config, you must specify the Amazon Resource Name (ARN) that AWS Lambda assigns to the function. Specify the ARN for the SourceIdentifier key. This key is part of the Source object, which is part of the ConfigRule object.

If you are adding an AWS managed Config rule, specify the rule's identifier for the SourceIdentifier key. To reference AWS managed Config rule identifiers, see About AWS Managed Config Rules.

For any new rule that you add, specify the ConfigRuleName in the ConfigRule object. Do not specify the ConfigRuleArn or the ConfigRuleId. These values are generated by AWS Config for new rules.

If you are updating a rule that you added previously, you can specify the rule by ConfigRuleName, ConfigRuleId, or ConfigRuleArn in the ConfigRule data type that you use in this request.

The maximum number of rules that AWS Config supports is 150.

For information about requesting a rule limit increase, see AWS Config Limits in the AWS General Reference Guide.

For more information about developing and using AWS Config rules, see Evaluating AWS Resource Configurations with AWS Config in the AWS Config Developer Guide.

fn put_configuration_aggregator(
&self,
input: PutConfigurationAggregatorRequest
) -> RusotoFuture<PutConfigurationAggregatorResponse, PutConfigurationAggregatorError>

Creates and updates the configuration aggregator with the selected source accounts and regions. The source account can be individual account(s) or an organization.

AWS Config should be enabled in source accounts and regions you want to aggregate.

If your source type is an organization, you must be signed in to the master account and all features must be enabled in your organization. AWS Config calls EnableAwsServiceAccess API to enable integration between AWS Config and AWS Organizations.

fn put_configuration_recorder(
&self,
input: PutConfigurationRecorderRequest
) -> RusotoFuture<(), PutConfigurationRecorderError>

Creates a new configuration recorder to record the selected resource configurations.

You can use this action to change the role roleARN or the recordingGroup of an existing recorder. To change the role, call the action on the existing configuration recorder and specify a role.

Currently, you can specify only one configuration recorder per region in your account.

If ConfigurationRecorder does not have the recordingGroup parameter specified, the default is to record all supported resource types.

fn put_delivery_channel(
&self,
input: PutDeliveryChannelRequest
) -> RusotoFuture<(), PutDeliveryChannelError>

Creates a delivery channel object to deliver configuration information to an Amazon S3 bucket and Amazon SNS topic.

Before you can create a delivery channel, you must create a configuration recorder.

You can use this action to change the Amazon S3 bucket or an Amazon SNS topic of the existing delivery channel. To change the Amazon S3 bucket or an Amazon SNS topic, call this action and specify the changed values for the S3 bucket and the SNS topic. If you specify a different value for either the S3 bucket or the SNS topic, this action will keep the existing value for the parameter that is not changed.

You can have only one delivery channel per region in your account.

fn put_evaluations(
&self,
input: PutEvaluationsRequest
) -> RusotoFuture<PutEvaluationsResponse, PutEvaluationsError>

Used by an AWS Lambda function to deliver evaluation results to AWS Config. This action is required in every AWS Lambda function that is invoked by an AWS Config rule.

fn put_remediation_configurations(
&self,
input: PutRemediationConfigurationsRequest
) -> RusotoFuture<PutRemediationConfigurationsResponse, PutRemediationConfigurationsError>

Adds or updates the remediation configuration with a specific AWS Config rule with the selected target or action. The API creates the RemediationConfiguration object for the AWS Config rule. The AWS Config rule must already exist for you to add a remediation configuration. The target (SSM document) must exist and have permissions to use the target.

fn put_retention_configuration(
&self,
input: PutRetentionConfigurationRequest
) -> RusotoFuture<PutRetentionConfigurationResponse, PutRetentionConfigurationError>

Creates and updates the retention configuration with details about retention period (number of days) that AWS Config stores your historical information. The API creates the RetentionConfiguration object and names the object as default. When you have a RetentionConfiguration object named default, calling the API modifies the default object.

Currently, AWS Config supports only one retention configuration per region in your account.

fn select_resource_config(
&self,
input: SelectResourceConfigRequest
) -> RusotoFuture<SelectResourceConfigResponse, SelectResourceConfigError>

Accepts a structured query language (SQL) SELECT command, performs the corresponding search, and returns resource configurations matching the properties.

For more information about query components, see the Query Components section in the AWS Config Developer Guide.

fn start_config_rules_evaluation(
&self,
input: StartConfigRulesEvaluationRequest
) -> RusotoFuture<StartConfigRulesEvaluationResponse, StartConfigRulesEvaluationError>

Runs an on-demand evaluation for the specified AWS Config rules against the last known configuration state of the resources. Use StartConfigRulesEvaluation when you want to test that a rule you updated is working as expected. StartConfigRulesEvaluation does not re-record the latest configuration state for your resources. It re-runs an evaluation against the last known state of your resources.

You can specify up to 25 AWS Config rules per request.

An existing StartConfigRulesEvaluation call for the specified rules must complete before you can call the API again. If you chose to have AWS Config stream to an Amazon SNS topic, you will receive a ConfigRuleEvaluationStarted notification when the evaluation starts.

You don't need to call the StartConfigRulesEvaluation API to run an evaluation for a new rule. When you create a rule, AWS Config evaluates your resources against the rule automatically.

The StartConfigRulesEvaluation API is useful if you want to run on-demand evaluations, such as the following example:

You have a custom rule that evaluates your IAM resources every 24 hours.

You update your Lambda function to add additional conditions to your rule.

Instead of waiting for the next periodic evaluation, you call the StartConfigRulesEvaluation API.

AWS Config invokes your Lambda function and evaluates your IAM resources.

Your custom rule will still run periodic evaluations every 24 hours.

fn start_configuration_recorder(
&self,
input: StartConfigurationRecorderRequest
) -> RusotoFuture<(), StartConfigurationRecorderError>

Starts recording configurations of the AWS resources you have selected to record in your AWS account.

You must have created at least one delivery channel to successfully start the configuration recorder.

fn start_remediation_execution(
&self,
input: StartRemediationExecutionRequest
) -> RusotoFuture<StartRemediationExecutionResponse, StartRemediationExecutionError>

Runs an on-demand remediation for the specified AWS Config rules against the last known remediation configuration. It runs an execution against the current state of your resources. Remediation execution is asynchronous.

You can specify up to 100 resource keys per request. An existing StartRemediationExecution call for the specified resource keys must complete before you can call the API again.

fn stop_configuration_recorder(
&self,
input: StopConfigurationRecorderRequest
) -> RusotoFuture<(), StopConfigurationRecorderError>

Stops recording configurations of the AWS resources you have selected to record in your AWS account.

fn tag_resource(
&self,
input: TagResourceRequest
) -> RusotoFuture<(), TagResourceError>

Associates the specified tags to a resource with the specified resourceArn. If existing tags on a resource are not specified in the request parameters, they are not changed. When a resource is deleted, the tags associated with that resource are deleted as well.

fn untag_resource(
&self,
input: UntagResourceRequest
) -> RusotoFuture<(), UntagResourceError>

Deletes specified tags from a resource.

Loading content...